requirements.txt
Flask==1.0.2
numpy==1.16.4
Pillow==6.1.0
torch==1.1.0
torchvision==0.3.0